refactor: extract and rename identity types
