+++ Applications/AdminUi/packages/admin_api_sdk/lib/src/endpoints/identity_endpoint.dart
@@ -7,12 +7,12 @@ import 'endpoint.dart';
 class IdentitiesEndpoint extends Endpoint {
   IdentitiesEndpoint(super._dio);
 
-  Future<ApiResponse<Identity>> getIdentity(
+  Future<ApiResponse<IdentityResponse>> getIdentity(
     String address,
   ) =>
       get(
         '/api/v1/Identities/$address',
-        transformer: Identity.fromJson,
+        transformer: IdentityResponse.fromJson,
       );
 
   Future<ApiResponse<void>> updateIdentity(
